admin

+++ app/views/menu_view.php
@@ -17,8 +17,10 @@
 		<?php 
 			
 			foreach ($data['goods'][0] as $key) {
-				$edit = '<a  href="/menu/edit/id/'.$key["id"].'" class="lsf symbol edit-icon" style="font-size: 30px;">edit</a>';
-				$delete = '<a  href="/menu/delete/id/'.$key["id"].'" class="lsf symbol delete-icon" style="font-size: 30px;">delete</a>';
+				if($data['authorized']){
+					$edit = '<a  href="/menu/edit/id/'.$key["id"].'" class="lsf symbol edit-icon" style="font-size: 30px;">edit</a>';
+					$delete = '<a  href="/menu/delete/id/'.$key["id"].'" class="lsf symbol delete-icon" style="font-size: 30px;">delete</a>';
+				}
 				$buy = 0;
 				$i = 0;
 				$display = 'none';
